requirements-chainlit-docker.txt
# Slim image for Chainlit + orchestrator + agents + RAG (retrieve).
# Avoids full requirements.txt (aiortc/PyAV/ffmpeg stack that breaks slim Docker builds).
#
# CPU-only PyTorch (smaller) — must stay at top for pip
--extra-index-url https://download.pytorch.org/whl/cpu

torch>=2.2.0,<3

chainlit>=1.3.0,<3.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0,<3

langgraph>=0.3.18
langchain-core>=0.3.0
langchain-openai>=0.3.9
langchain-chroma>=0.2.6
langchain-community>=0.3.20

chromadb>=0.5.0
sentence-transformers>=3.0.0

httpx>=0.27.0
openai>=1.40.0
tiktoken>=0.7.0
numpy>=1.26.0
